My new test, and cleanup of Gaurav's login test
This test checks to make sure that, when we successfully authenticate, the environment is set up correctly by pam_google.  Also, I cleaned up some stuff in the other existing authN test to make it stop hanging autotest.

Review URL: http://chromereview.prom.corp.google.com/1187079

git-svn-id: svn://chrome-svn/chromeos/trunk@274 06c00378-0e64-4dae-be16-12b19f9950a1

+++ src/platform/testing/systemtests/login_AuthenticationEnvVarsTests/login_AuthenticationEnvVarsTests.py
@@ -8,11 +8,7 @@ import subprocess
 from autotest_lib.client.bin import test
 from autotest_lib.client.common_lib import error
 
-class login_AuthenticationTests(test.test):
-  """
-  Tests if we can successfully authenticate to google using credentials listed
-  in authentication-test.dat.
-  """
+class login_AuthenticationEnvVarsTests(test.test):
 
   version = 1
   def run_once(self):
@@ -21,29 +17,27 @@ class login_AuthenticationTests(test.test):
     # than doing an eval()
     auth_test_data = eval(fd.read())
     try:
-      success = self.RunAuthenticationTests(auth_test_data)
-      if not success:
-        raise error.TestFail("Failed to Authenticate")
+      self.RunAuthenticationTests(auth_test_data)
     except error.TestFail, tfe:
       raise tfe
     except Exception, e:
       raise error.TestError(e)
 
-  def RunAuthenticationTests(self,auth_data):
+  def RunAuthenticationEnvVarsTests(self,auth_data):
     pipe_path = '/tmp/cookie_pipe'
     for username, properties in auth_data.iteritems():
-      login_cmd = os.path.join(self.bindir, "login_AuthenticationTest")
+      login_cmd = os.path.join(self.bindir, "login_AuthenticationEnvVarsTest")
       password = properties.get("password", None)
       cmdline = [login_cmd, username, password]
-      return_code = subprocess.call(cmdline)
-      if return_code == 255:
-        return False
-      if return_code == 1:
-        raise error.TestError("Test failed")
-      if os.path.exists(pipe_path):
+      return_code = subprocess.call(cmdline, shell=False)
+      if return_code == 0 and os.path.exists(pipe_path):
         """if we successfully authenticated, there's a process waiting to send
         us cookies on a named pipe.  We need to read them."""
         for line in file(pipe_path):
           pass
         os.unlink(pipe_path)
+      elif return_code == 255:
+        raise error.TestFail("Authentication or environment setup failed.")
+      else:
+        raise error.TestError("Unknown test error.")
     return True
